Updated url parameter parsing

+++ checkIn/server/db.js
@@ -18,7 +18,7 @@ const Pricing = sequelize.define('pricing', {
     autoIncrement: true,
   },
   base_price: {
-    type: DECIMAL(18, 2),
+    type: INTEGER,
     allowNull: false,
   },
   cleaning_fee: {
